Force consistent ordering of dates containing NULLs
Because database servers differ on how they sort NULL values, the sort
order for articles was changed, via COALESCE, so that a NULL date will
be treated as being less than (i.e., older than) any non-NULL date.
This fixes test_should_search_article_by_section so that it passes with
MySQL, SQLite, and PostgreSQL.

[I'm merging this patch because I'm quite fond of PostgreSQL, and
because it has been tested with most of the other databases we support.
If this breaks your database, please let me know. -Eric]